+++ server/test/integration_test.sh
@@ -1,16 +1,16 @@
 #!/bin/bash
 
-cd ./util/test-files
+cd ./test
 source ./setenv.sh
 docker compose up -d 
 echo 'Waiting for database to be ready...'
 ./wait-for-it.sh "${DATABASE_URL}" -- echo 'Database is ready!'
 
-cd ../../../db
+cd ../../db
 npx prisma migrate dev
 
 cd ../server
-vitest -c ./util/test-files/vitest.config.integration.ts
+vitest -c ./test/vitest.config.integration.ts
 
-cd ./util/test-files
+cd ./test
 docker compose down
