+++ server/test/setenv.sh
