+++ server/test/vitest.config.integration.ts
@@ -2,6 +2,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['./persistence/**/*.test.ts'],
+    include: ['./test/persistence/*.test.ts'],
   },
 });
